+++ test/integration/tests/hash.sh
@@ -24,9 +24,9 @@ cleanup "no-shut-down"
 
 echo "T0naX0u123abc" > $hash_in_file
 
-# Test with ticket and hash output files and verify that the output hash
+# Test with ticket and hash output files (binary) and verify that the output hash
 # is correct. Ticket is not stable and changes run to run, don't verify it.
-tpm2_hash -C e -g sha1 -o $hash_out_file -t $ticket_file $hash_in_file 1>/dev/null
+tpm2_hash -C e -g sha1 -o $hash_out_file -t $ticket_file $hash_in_file
 
 expected=`sha1sum $hash_in_file | awk '{print $1}'`
 actual=`cat $hash_out_file | xxd -p -c 20`
@@ -36,9 +36,9 @@ test "$expected" == "$actual"
 cleanup "no-shut-down"
 
 # Test platform hierarchy with multiple files and verify output against sha256sum
-# Test a file redirection as well.
+# Test a file redirection as well. Output files are binary.
 echo "T0naX0u123abc" > $hash_in_file
-tpm2_hash -C p -g sha256 -Q -o $hash_out_file -t $ticket_file < $hash_in_file
+tpm2_hash -C p -g sha256 -o $hash_out_file -t $ticket_file < $hash_in_file
 
 expected=`sha256sum $hash_in_file | awk '{print $1}'`
 actual=`cat $hash_out_file | xxd -p -c 32`
@@ -48,8 +48,8 @@ test "$expected" == "$actual"
 cleanup "no-shut-down"
 
 # Test stdout output as well as no options.
-# Validate that hash outputs are as expected.
-tpm_hash_val=`echo 1234 | tpm2_hash -C n | tee $out | grep sha1 | cut -d\: -f 2-2 | tr -d '[:space:]'`
+# Validate that hash outputs are in hex as expected.
+tpm_hash_val=`echo 1234 | tpm2_hash -C n --hex`
 sha1sum_val=`echo 1234 | sha1sum  | cut -d\  -f 1-2 | tr -d '[:space:]'`
 if [ "$tpm_hash_val" != "$sha1sum_val" ]; then
   echo "Expected tpm and sha1sum to produce same hashes."
@@ -59,12 +59,9 @@ if [ "$tpm_hash_val" != "$sha1sum_val" ]; then
   exit 1
 fi
 
-# Verify the yaml output
-yaml_verify $out
-
 # Test a file that cannot be done in 1 update call. The tpm works on a 1024 block size.
 dd if=/dev/urandom of=$hash_in_file bs=2093 count=1 2>/dev/null
-tpm_hash_val=`tpm2_hash $hash_in_file | grep sha1 | cut -d\: -f 2-2 | tr -d '[:space:]'`
+tpm_hash_val=`tpm2_hash --hex $hash_in_file`
 sha1sum_val=`sha1sum $hash_in_file | cut -d\  -f 1-2 | tr -d '[:space:]'`
 if [ "$tpm_hash_val" != "$sha1sum_val" ]; then
   echo "Expected tpm and sha1sum to produce same hashes"